refactor(frontend): centralize shared page data types in src/types
Move duplicated tournament, player, match hub, admin, and leaderboard
shapes out of page files into shared types and wire pages to import them.

Closes #366
