+++ frontend/src/app/play/party/page.tsx
@@ -18,17 +18,10 @@ import PartyManager from '@/components/game/PartyManager';
 import CountdownTimer from '@/components/game/CountdownTimer';
 import ChatPanel from '@/components/game/ChatPanel';
 import PlayerList from '@/components/game/PlayerList';
+import type { PartyPlayer } from '@/types/player';
 
 const MAX_PARTY_SIZE = 4;
 const COUNTDOWN_SECONDS = 15;
-
-interface PartyPlayer {
-    id: string;
-    username: string;
-    avatar?: string;
-    isReady: boolean;
-    isHost: boolean;
-}
 
 export default function PlayPartyPage() {
     const router = useRouter();
