+++ frontend/src/data/matchHub.ts
@@ -1,47 +1,6 @@
-import { BracketMatch, ScoreReport } from "@/types/bracket";
+import type { MatchHubDetails } from "@/types/match";
 
-export interface MatchHubEvent {
-  id: string;
-  type: "status" | "score" | "alert" | "report";
-  message: string;
-  createdAt: string;
-}
-
-export interface MatchHubPlayerSnapshot {
-  id: string;
-  username: string;
-  avatar?: string;
-  elo: number;
-  region: string;
-  seed: number;
-  record: string;
-  stats: Array<{ label: string; value: string }>;
-}
-
-export interface MatchHubDetails {
-  id: string;
-  tournamentId: string;
-  tournamentName: string;
-  gameType: string;
-  bracketFormat: "single_elimination" | "double_elimination";
-  roundLabel: string;
-  arenaLabel: string;
-  status: BracketMatch["status"];
-  bestOf: number;
-  scheduledTime: string;
-  startedAt?: string;
-  streamTitle?: string;
-  prizePool: number;
-  player1: MatchHubPlayerSnapshot;
-  player2: MatchHubPlayerSnapshot;
-  scorePlayer1: number;
-  scorePlayer2: number;
-  winnerId?: string;
-  notes: string;
-  reports: ScoreReport[];
-  feed: MatchHubEvent[];
-  canDisputeUntil: string;
-}
+export type { MatchHubDetails, MatchHubEvent, MatchHubPlayerSnapshot } from "@/types/match";
 
 export const matchHubDetails: Record<string, MatchHubDetails> = {
   "1-match-13": {
