+++ frontend/src/types/index.ts
@@ -1,11 +1,15 @@
 // Export all types
 export * from './achievement';
+export * from './admin';
 export * from './bracket';
 export * from './leaderboard';
 export * from './match';
 export * from './notification';
+export * from './player';
+export * from './profile';
 export * from './social';
 export * from './tournament';
+export * from './transaction';
 export * from './user';
 
 // Common API response types
