+++ frontend/src/types/tournament.ts
@@ -48,6 +48,26 @@ export interface CreateTournamentRequest {
   startTime: string;
 }
 
+export type CreateTournamentWizardStep = 1 | 2 | 3 | 4 | "preview" | "success";
+
+export interface CreateTournamentFormData {
+  name: string;
+  gameType: string;
+  description: string;
+  tournamentType: TournamentType;
+  matchFormat: string;
+  rules: string;
+  entryFee: number;
+  prizePool: number;
+  prizeDistribution: string;
+  visibility: TournamentVisibility;
+  maxParticipants: number;
+  registrationOpenDate: string;
+  registrationCloseDate: string;
+  startDate: string;
+  endDate: string;
+}
+
 export interface TournamentFilters {
   gameType?: string;
   status?: TournamentStatus;
