+++ frontend/src/types/transaction.ts
@@ -1,0 +1,11 @@
+export type {
+  TxDirection,
+  TxHistoryItem,
+  TxKind,
+  TxMeta,
+  TxPhase,
+  TxStatus,
+} from "@/lib/wallet/types";
+
+/** Wallet transaction entry (deposit / withdraw) shown in history UIs. */
+export type { TxHistoryItem as Transaction } from "@/lib/wallet/types";
